requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==38.0.1
cssselect==1.1.0
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.2.0
filelock==3.8.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
pluggy==1.0.0
Protego==0.2.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.1
tomli==2.0.1
Twisted==22.8.0
typing-extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
zope.interface==5.4.0
redis==4.3.4
django-redis==5.2.0
django-rq==2.5.1
python-telegram-bot==20.0a4
gunicorn==20.1.0

flake8>=5.0.4
isort>=5.10.1
black>=22.10.0